requirements.txt
Django==1.5.1
South==0.8.2
django-autocomplete-light==1.4.3
django-crispy-forms==1.4.0
django-registration==1.0
wsgiref==0.1.2
